+++ src/ung_wishlist/AccountScreen.java
@@ -1,7 +1,6 @@
 package ung_wishlist;
 
 import javax.swing.*;
-
 import java.awt.BorderLayout;
 import java.awt.GridLayout;
 import java.sql.ResultSet;
@@ -13,6 +12,10 @@ import java.awt.GridBagConstraints;
 import java.awt.Insets;
 import java.awt.FlowLayout;
 import java.awt.CardLayout;
+import java.awt.event.ActionListener;
+import java.awt.event.ActionEvent;
+import java.beans.PropertyChangeListener;
+import java.beans.PropertyChangeEvent;
 
 // Class purpose: 
 // Account options with gifts list here and buttons to create a new one
@@ -22,6 +25,7 @@ import java.awt.CardLayout;
 public class AccountScreen extends JPanel{
 	private MainFrame mainFrame;  
 	private User currentUser; 
+	ResultSet resultSet = null;
 	
 	// Passed currentUser class from LoginPanel stores
 	public AccountScreen(MainFrame mainFrame, User currentUser) {
@@ -36,6 +40,8 @@ public class AccountScreen extends JPanel{
 		JPanel rightPanel = new JPanel(new BorderLayout());
 		setLayout(new BorderLayout(0, 0));
 		JSplitPane splitPane = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT, leftPanel, rightPanel);
+		
+
 		splitPane.setDividerLocation(400);
 		splitPane.setResizeWeight(1.0);
 		add(splitPane);
@@ -49,6 +55,9 @@ public class AccountScreen extends JPanel{
 		JLabel usernameLabel = new JLabel(currentUser.getUsername());
 		JLabel emailLabel = new JLabel(currentUser.getEmail());
 		JButton logoutButton = new JButton("Log Out");
+		
+
+		
 		userInfoPanel.add(logoutButton);
 		userInfoPanel.add(nameLabel);
 		userInfoPanel.add(usernameLabel);
@@ -61,6 +70,7 @@ public class AccountScreen extends JPanel{
 		JButton createListButton = new JButton("Create List");
 		JButton editListButton = new JButton("Edit List");
 		JButton deleteListButton = new JButton("Delete List");
+
 		
 		
 		buttonPanel.add(createListButton);
@@ -78,6 +88,8 @@ public class AccountScreen extends JPanel{
 		
 		leftPanel.add(searchPanel);
 		
+		// User Search 
+		
 		JLabel lblNewLabel = new JLabel("User Search");
 		GridBagConstraints gbc_lblNewLabel = new GridBagConstraints();
 		gbc_lblNewLabel.insets = new Insets(0, 0, 5, 5);
@@ -94,6 +106,7 @@ public class AccountScreen extends JPanel{
 		gbc_searchField.gridy = 6;
 		searchPanel.add(searchField, gbc_searchField);
 		JButton searchButton = new JButton("Search");
+
 		GridBagConstraints gbc_searchButton = new GridBagConstraints();
 		gbc_searchButton.insets = new Insets(0, 0, 0, 5);
 		gbc_searchButton.fill = GridBagConstraints.BOTH;
@@ -103,19 +116,79 @@ public class AccountScreen extends JPanel{
 		
 		// Right panel components
 
+
 		
+		JLabel listLabel = new JLabel("My Lists");
+		listLabel.setHorizontalAlignment(SwingConstants.CENTER);
+		rightPanel.add(listLabel, BorderLayout.NORTH);
 		
+		JLabel noListsLabel = new JLabel("");
+		noListsLabel.setHorizontalAlignment(SwingConstants.CENTER);
+		noListsLabel.setEnabled(false);
+		rightPanel.add(noListsLabel, BorderLayout.SOUTH);
+		
+		JList<String> list = new JList<>();
+		Authentication.getUserLists(currentUser.getId(), list);
+		
+		if (list.getModel().getSize() == 0) {
+			noListsLabel.setText("You don't have a list!");
+			noListsLabel.setEnabled(true);
+		}
+		
+		rightPanel.add(new JScrollPane(list), BorderLayout.CENTER);
+		
+
+		
+		// Button actions
+		
+		// Logout user
+		logoutButton.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				currentUser.logout();
+				mainFrame.showLoginPanel();
+			}
+		});
+		
+		// Create list
+		createListButton.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				String name = JOptionPane.showInputDialog("Enter name of list: ");
+				if (Authentication.checkListExists(name)) {
+					JOptionPane.showMessageDialog(null, "List already exits!");
+				} else {
+					Authentication.createList(currentUser.getId(), name);
+					Authentication.getUserLists(currentUser.getId(), list);
+				}
+			}
+		});
+		
+		// Edit list
+		editListButton.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				JOptionPane.showMessageDialog(null, "Not functioning");
+			}
+		});
+		// Delete List
+		deleteListButton.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				String listName = list.getSelectedValue();
+				if (listName != null) {
+					Authentication.deleteList(currentUser.getId(), listName);
+					Authentication.getUserLists(currentUser.getId(), list);
+				} else {
+					JOptionPane.showMessageDialog(null, "No list selected.");
+				}
+				
+			}
+		});		
+		
+		// Search Users
+		searchButton.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				JOptionPane.showMessageDialog(null, "Not functioning");
+			}
+		});
 	}
-	
-	
-	
-	// Methods
-	
-		// Logout method
-			// Clear store user information in User class 
-			// Return to Login screen
-	
-		// Change Account Details (pass, user, name, email)
 	
 	
 }