+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Tyler Stanton 3/24/2024
  - Added Create Account Screen by Sam
  - Creating account works with database.
  
- Tyler Stanton 3/29/2024
+Tyler Stanton 3/29/2024
  - Login screen checks database for login
  - Once logged in, info saved in User class
+ 
+Tyler Stanton 4/08/2024
+ - Added Account screen that displays user's information, contains logout, create list, edit list, delete list buttons.
+ - Users can now create lists
+ - Users can now delete lists